+++ internal/utils/fileedit.go
@@ -33,9 +33,10 @@ import (
 )
 
 const (
-	CONFIG_PREFIX = "# BEGIN_AWS_SSO_CLI"
-	CONFIG_SUFFIX = "# END_AWS_SSO_CLI"
-	FILE_TEMPLATE = "%s\n%s\n%s\n"
+	DEFAULT_SSO_NAME = "Default" // DefaultSSO per cmd/aws-sso/main.go
+	CONFIG_PREFIX    = "# BEGIN_AWS_SSO_CLI"
+	CONFIG_SUFFIX    = "# END_AWS_SSO_CLI"
+	FILE_TEMPLATE    = "%s\n%s\n%s\n"
 )
 
 type FileEdit struct {
@@ -62,11 +63,17 @@ func GenerateSource(fileTemplate string, vars interface{}) ([]byte, error) {
 	return buf.Bytes(), nil
 }
 
-func NewFileEdit(fileTemplate string, vars interface{}) (*FileEdit, error) {
+func NewFileEdit(fileTemplate, ssoName string, vars interface{}) (*FileEdit, error) {
 	var t string
 
 	if fileTemplate != "" {
-		t = fmt.Sprintf(FILE_TEMPLATE, CONFIG_PREFIX, fileTemplate, CONFIG_SUFFIX)
+		prefix := CONFIG_PREFIX
+		suffix := CONFIG_SUFFIX
+		if ssoName != "" && ssoName != DEFAULT_SSO_NAME {
+			prefix = fmt.Sprintf("%s_%s", CONFIG_PREFIX, ssoName)
+			suffix = fmt.Sprintf("%s_%s", CONFIG_SUFFIX, ssoName)
+		}
+		t = fmt.Sprintf(FILE_TEMPLATE, prefix, fileTemplate, suffix)
 	}
 	templ, err := template.New("template").Parse(t)
 	if err != nil {
@@ -84,8 +91,9 @@ func NewFileEdit(fileTemplate string, vars interface{}) (*FileEdit, error) {
 var diffWriter io.Writer = os.Stdout
 
 // UpdateConfig does all the heavy lifting of updating (or creating) the file
-// and optionally providing a diff for user to approve/view
-func (f *FileEdit) UpdateConfig(printDiff, force bool, configFile string) error {
+// and optionally providing a diff for user to approve/view.  Returns true if
+// changes were made, false if no changes were made, or an error if something happened
+func (f *FileEdit) UpdateConfig(printDiff, force bool, configFile string) (bool, error) {
 	inputBytes, err := os.ReadFile(configFile)
 	if err != nil {
 		inputBytes = []byte{}
@@ -93,7 +101,7 @@ func (f *FileEdit) UpdateConfig(printDiff, force bool, configFile string) error 
 
 	outputBytes, err := f.GenerateNewFile(configFile)
 	if err != nil {
-		return err
+		return false, err
 	}
 
 	newFile := fmt.Sprintf("%s.new", configFile)
@@ -103,22 +111,22 @@ func (f *FileEdit) UpdateConfig(printDiff, force bool, configFile string) error 
 	if len(diff) == 0 {
 		// do nothing if there is no diff
 		log.Infof("no changes made to %s", configFile)
-		return nil
+		return false, nil
 	}
 
 	if !force {
 		approved, err := prompt(configFile, diff)
 		if err != nil {
-			return err
+			return false, err
 		}
 		if !approved {
-			return nil
+			return false, nil
 		}
 	} else if printDiff {
 		fmt.Fprintf(diffWriter, "%s", diff)
 	}
 
-	return os.WriteFile(configFile, outputBytes, 0600)
+	return true, os.WriteFile(configFile, outputBytes, 0600)
 }
 
 // GenerateNewFile generates the contents of a new config file